Curso_Python/14_DataBase_MySQL/PersonasDB/requirements.txt
mysql_connector_python>=9.3.0
